Nettoyage : éclate les calculateurs par niveau
