+++ spec/models/restitution/calculateur_scores_niveau1_spec.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+require 'rails_helper'
+
+describe Restitution::CalculateurScoresNiveau1 do
+  let(:calculateur_scores_niveau2) { double }
+
+  let(:calculateur_scores_niveau1) do
+    Restitution::CalculateurScoresNiveau1.new(calculateur_scores_niveau2)
+  end
+
+  describe '#scores_niveau1' do
+    it 'à partir des score standardisés de niveau 2' do
+      allow(calculateur_scores_niveau2).to receive(:scores_niveau2_standardises)
+        .and_return(score_numeratie: 1,
+                    score_ccf: 2,
+                    score_syntaxe_orthographe: 3,
+                    score_memorisation: 4)
+      expect(calculateur_scores_niveau1.scores_niveau1)
+        .to eq(litteratie: (2 + 3 + 4) / 3.0, numeratie: 1)
+    end
+
+    it 'avec une metrique manquante' do
+      allow(calculateur_scores_niveau2).to receive(:scores_niveau2_standardises)
+        .and_return(score_numeratie: 1,
+                    score_ccf: nil,
+                    score_syntaxe_orthographe: 3,
+                    score_memorisation: 4)
+      expect(calculateur_scores_niveau1.scores_niveau1)
+        .to eq(litteratie: (3 + 4) / 2.0, numeratie: 1)
+    end
+  end
+end
