+++ spec/models/restitution/globale_spec.rb
@@ -146,72 +146,4 @@ describe Restitution::Globale do
       it { expect(restitution_globale.niveaux_competences).to eq([]) }
     end
   end
-
-  describe '#calcul_scores_niveau2_metriques' do
-    let(:standardisateur) { double }
-    let(:situation_id) { double }
-    before do
-      allow(restitution_globale).to receive(:standardisateurs_niveau3)
-        .and_return({ situation_id => standardisateur })
-      allow(standardisateur).to receive(:standardise).and_return(nil)
-    end
-
-    context 'pas de restitution' do
-      let(:restitutions) { [] }
-      it do
-        expect(restitution_globale.scores_niveau2).to eq({})
-      end
-    end
-
-    context 'une restitution avec un score' do
-      let(:partie) { double(situation_id: situation_id) }
-      let(:restitutions) { [double(partie: partie)] }
-      it do
-        allow(partie).to receive(:metriques).and_return({ 'score_ccf' => 110 })
-        allow(standardisateur).to receive(:standardise).with(:score_ccf, 110).and_return(1.1)
-        expect(restitution_globale.scores_niveau2).to eq(score_ccf: 1.1)
-      end
-    end
-
-    context 'fait la moyenne des scores_niveau2 de restitution' do
-      let(:partie1) { double(situation_id: situation_id) }
-      let(:partie2) { double(situation_id: situation_id) }
-      let(:restitutions) { [double(partie: partie1), double(partie: partie2)] }
-      it do
-        allow(partie1).to receive(:metriques).and_return({ 'score_ccf' => 110 })
-        allow(partie2).to receive(:metriques).and_return({ 'score_ccf' => 120 })
-
-        allow(standardisateur).to receive(:standardise).with(:score_ccf, 110).and_return(1.1)
-        allow(standardisateur).to receive(:standardise).with(:score_ccf, 120).and_return(1.2)
-
-        expect(restitution_globale.scores_niveau2).to eq(score_ccf: 1.15)
-      end
-    end
-
-    context 'sépare les scores_niveau2 des compétences différentes' do
-      let(:partie1) { double(situation_id: situation_id) }
-      let(:partie2) { double(situation_id: situation_id) }
-      let(:restitutions) { [double(partie: partie1), double(partie: partie2)] }
-
-      it do
-        allow(partie1).to receive(:metriques)
-          .and_return({ 'score_ccf' => 110, 'score_memorisation' => 120 })
-        allow(partie2).to receive(:metriques).and_return({ 'score_numeratie' => 130 })
-
-        allow(standardisateur).to receive(:standardise)
-          .with(:score_ccf, 110)
-          .and_return(1.1)
-        allow(standardisateur).to receive(:standardise)
-          .with(:score_memorisation, 120)
-          .and_return(1.2)
-        allow(standardisateur).to receive(:standardise)
-          .with(:score_numeratie, 130)
-          .and_return(1.3)
-
-        expect(restitution_globale.scores_niveau2).to eq(score_ccf: 1.1,
-                                                         score_memorisation: 1.2,
-                                                         score_numeratie: 1.3)
-      end
-    end
-  end
 end
